// comercial-jhoel-api/src/modules/landing-backgrounds/presentation/dtos/landing-background.response.dto.ts
export class LandingBackgroundResponseDto {
  id: string;
  name: string;
  sectionKey: string;
  opacity: number;
  overlay: string;
  position: string;
  size: string;
  depthEffect: string;
  parallax: string;
  movement: string;
  isActive: boolean;
  hasImage: boolean;
  createdAt: Date;
  updatedAt: Date;
  createdBy: string;
  createdByUsername: string;
  updatedBy: string | null;
  updatedByUsername: string | null;
}